Salary Changing

+++ HR/DTO/SalaryReportDto.cs
@@ -12,10 +12,9 @@
         public int extraHours { get; set; }
         public int dedectionHours { get; set; }
         public int extraTimebeforSetting { get; set; }
-
         public int discountTimebeforSetting { get; set; }
-        public int totalExtra { get; set; }
-        public int totalDiscount { get; set; }
+        public double totalExtra { get; set; }
+        public double totalDiscount { get; set; }
         public double totalNetSalary { get; set; }
     }
 }
